Make time values longs instead of ints.

+++ structs/stn.hpp
@@ -4,7 +4,7 @@
 
 class Stn {
 public:
-	typedef int Time;
+	typedef long Time;
 
 	Stn(unsigned int);
